feat/ui/upload-image: the user register form and add car form is modifyied now we can send image to backend
