+++ frontend/src/index.js
@@ -11,7 +11,6 @@ import CarDetail from "./Components/CarDetail";
 import { QueryClient, QueryClientProvider } from "react-query";
 import { ReactQueryDevtools } from "react-query/devtools";
 import Users from "./Components/Delete";
-import Sign from "./Components/Sign";
 import Regisrer from "./Components/user/Regisrer";
 import Logout from "./Components/user/Logout";
 import UpdateCar from "./Components/UpdateCar";
@@ -52,10 +51,6 @@ const router = createBrowserRouter([
   {
     path: "/del/",
     element: <Users />,
-  },
-  {
-    path: "/for/",
-    element: <Sign />,
   },
   {
     path: "/reg/",
